+++ featuretools/primitives/transform/datetime/is_weekend.py
@@ -1,0 +1,41 @@
+import holidays
+import numpy as np
+import pandas as pd
+from woodwork.column_schema import ColumnSchema
+from woodwork.logical_types import (
+    AgeFractional,
+    BooleanNullable,
+    Categorical,
+    Datetime,
+    Ordinal,
+)
+
+from featuretools.primitives.core.transform_primitive import TransformPrimitive
+from featuretools.primitives.utils import HolidayUtil
+from featuretools.utils import convert_time_units
+from featuretools.utils.gen_utils import Library
+
+class IsWeekend(TransformPrimitive):
+    """Determines if a date falls on a weekend.
+
+    Examples:
+        >>> from datetime import datetime
+        >>> dates = [datetime(2019, 3, 1),
+        ...          datetime(2019, 6, 17, 11, 10, 50),
+        ...          datetime(2019, 11, 30, 19, 45, 15)]
+        >>> is_weekend = IsWeekend()
+        >>> is_weekend(dates).tolist()
+        [False, False, True]
+    """
+
+    name = "is_weekend"
+    input_types = [ColumnSchema(logical_type=Datetime)]
+    return_type = ColumnSchema(logical_type=BooleanNullable)
+    compatibility = [Library.PANDAS, Library.DASK, Library.SPARK]
+    description_template = "whether {} occurred on a weekend"
+
+    def get_function(self):
+        def is_weekend(vals):
+            return vals.dt.weekday > 4
+
+        return is_weekend
