+++ featuretools/primitives/transform/datetime/time_since.py
@@ -1,0 +1,64 @@
+import holidays
+import numpy as np
+import pandas as pd
+from woodwork.column_schema import ColumnSchema
+from woodwork.logical_types import (
+    AgeFractional,
+    BooleanNullable,
+    Categorical,
+    Datetime,
+    Ordinal,
+)
+
+from featuretools.primitives.core.transform_primitive import TransformPrimitive
+from featuretools.primitives.utils import HolidayUtil
+from featuretools.utils import convert_time_units
+from featuretools.utils.gen_utils import Library
+
+class TimeSince(TransformPrimitive):
+    """Calculates time from a value to a specified cutoff datetime.
+
+    Args:
+        unit (str): Defines the unit of time to count from.
+            Defaults to Seconds. Acceptable values:
+            years, months, days, hours, minutes, seconds, milliseconds, nanoseconds
+
+    Examples:
+        >>> from datetime import datetime
+        >>> time_since = TimeSince()
+        >>> times = [datetime(2019, 3, 1, 0, 0, 0, 1),
+        ...          datetime(2019, 3, 1, 0, 0, 1, 0),
+        ...          datetime(2019, 3, 1, 0, 2, 0, 0)]
+        >>> cutoff_time = datetime(2019, 3, 1, 0, 0, 0, 0)
+        >>> values = time_since(times, time=cutoff_time)
+        >>> list(map(int, values))
+        [0, -1, -120]
+
+        Change output to nanoseconds
+
+        >>> from datetime import datetime
+        >>> time_since_nano = TimeSince(unit='nanoseconds')
+        >>> times = [datetime(2019, 3, 1, 0, 0, 0, 1),
+        ...          datetime(2019, 3, 1, 0, 0, 1, 0),
+        ...          datetime(2019, 3, 1, 0, 2, 0, 0)]
+        >>> cutoff_time = datetime(2019, 3, 1, 0, 0, 0, 0)
+        >>> values = time_since_nano(times, time=cutoff_time)
+        >>> list(map(lambda x: int(round(x)), values))
+        [-1000, -1000000000, -120000000000]
+    """
+
+    name = "time_since"
+    input_types = [ColumnSchema(logical_type=Datetime)]
+    return_type = ColumnSchema(semantic_tags={"numeric"})
+    uses_calc_time = True
+    compatibility = [Library.PANDAS, Library.DASK]
+    description_template = "the time from {} to the cutoff time"
+
+    def __init__(self, unit="seconds"):
+        self.unit = unit.lower()
+
+    def get_function(self):
+        def pd_time_since(array, time):
+            return convert_time_units((time - array).dt.total_seconds(), self.unit)
+
+        return pd_time_since
