+++ featuretools/primitives/transform/datetime/weekday.py
@@ -1,0 +1,48 @@
+import holidays
+import numpy as np
+import pandas as pd
+from woodwork.column_schema import ColumnSchema
+from woodwork.logical_types import (
+    AgeFractional,
+    BooleanNullable,
+    Categorical,
+    Datetime,
+    Ordinal,
+)
+
+from featuretools.primitives.core.transform_primitive import TransformPrimitive
+from featuretools.primitives.utils import HolidayUtil
+from featuretools.utils import convert_time_units
+from featuretools.utils.gen_utils import Library
+
+class Weekday(TransformPrimitive):
+    """Determines the day of the week from a datetime.
+
+    Description:
+        Returns the day of the week from a datetime value. Weeks
+        start on Monday (day 0) and run through Sunday (day 6).
+
+    Examples:
+        >>> from datetime import datetime
+        >>> dates = [datetime(2019, 3, 1),
+        ...          datetime(2019, 6, 17, 11, 10, 50),
+        ...          datetime(2019, 11, 30, 19, 45, 15)]
+        >>> weekday = Weekday()
+        >>> weekday(dates).tolist()
+        [4, 0, 5]
+    """
+
+    name = "weekday"
+    input_types = [ColumnSchema(logical_type=Datetime)]
+    return_type = ColumnSchema(
+        logical_type=Ordinal(order=list(range(7))),
+        semantic_tags={"category"},
+    )
+    compatibility = [Library.PANDAS, Library.DASK, Library.SPARK]
+    description_template = "the day of the week of {}"
+
+    def get_function(self):
+        def weekday(vals):
+            return vals.dt.weekday
+
+        return weekday
